+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `filelog`, `journald`, `syslog`, `tcplog`, `udplog`: Remove `preserve_to` field from sub-parsers (#9331)
 - `kafkametricsreceiver`: instrumentation name updated from `otelcol/kafkametrics` to `otelcol/kafkametricsreceiver` (#9406)
 - `kubeletstatsreceiver`: instrumentation name updated from `kubeletstats` to `otelcol/kubeletstatsreceiver` (#9400)
+- `datadogexporter`: Remove `GetHostTags` method from `TagsConfig` struct (#9423)
 
 ### 🚩 Deprecations 🚩